requirements.txt
requests>=2.31
feedparser>=6.0
python-dateutil>=2.8
pandas>=2.1
